Throw appropriate exceptions from controller

+++ src/main/java/my/projects/passwordcards/rest/CredentialController.java
@@ -3,8 +3,10 @@ package my.projects.passwordcards.rest;
 import my.projects.passwordcards.model.Credential;
 import my.projects.passwordcards.service.CredentialService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -12,6 +14,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.servlet.ModelAndView;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 
 @Controller
 public class CredentialController {
@@ -38,6 +41,9 @@ public class CredentialController {
 
     @RequestMapping(value = "/save", method = RequestMethod.POST)
     public String saveCredential(@ModelAttribute("credential") Credential credential) {
+        if(null == credential || !StringUtils.hasLength(credential.getWebsite())) {
+            throw new IllegalArgumentException("Credential object cannot be null");
+        }
         credentialService.save(credential);
         // go back to the main page - index.html
         return "redirect:/";
@@ -46,8 +52,13 @@ public class CredentialController {
 
     @RequestMapping("/edit/{id}")
     public ModelAndView showEditCredentialPage(@PathVariable(name = "id") int id) {
+        if(0 >= id) {
+            throw new IllegalArgumentException("Illegal ID value: " + id);
+        }
         ModelAndView mav = new ModelAndView("editCredential");
+        // NoSuchElementException will be thrown and handled by ErrorHandler class
         Credential cr = credentialService.get(id);
+
         mav.addObject("credential", cr);
 
         return mav;
@@ -55,7 +66,16 @@ public class CredentialController {
 
     @RequestMapping("/delete/{id}")
     public String deleteCredential(@PathVariable(name = "id") int id) {
-        credentialService.delete(id);
+        if(0 >= id) {
+            throw new IllegalArgumentException("Illegal ID value: " + id);
+        }
+        try {
+            credentialService.delete(id);
+        } catch (EmptyResultDataAccessException e) {
+            System.err.println("Element not found. ID = " + id);
+            throw e;
+        }
+        // all other exceptions will be handles by ErrorHandler class
         return "redirect:/";
     }
 }
